pyproject.toml
[project]
name = "ai-chess-coach-rpg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain",
    "langgraph",
    "openai",
    "tiktoken",
    "python-chess",
    "stockfish",
    "rich",
    "typing_extensions",
    "dotenv"
]
